move green's function and follow PEP8
